Use generalized l2 sync worker
