+++ crates/common/Cargo.toml
@@ -39,12 +39,16 @@ sov-db = { path = "../sovereign-sdk/full-node/db/sov-db" }
 sov-ledger-rpc = { path = "../sovereign-sdk/full-node/sov-ledger-rpc", features = ["client"] }
 sov-mock-da = { path = "../sovereign-sdk/adapters/mock-da" }
 sov-modules-api = { path = "../sovereign-sdk/module-system/sov-modules-api" }
+sov-modules-stf-blueprint = { path = "../sovereign-sdk/module-system/sov-modules-stf-blueprint", features = ["native"] }
+sov-prover-storage-manager = { path = "../sovereign-sdk/full-node/sov-prover-storage-manager" }
 sov-rollup-interface = { path = "../sovereign-sdk/rollup-interface" }
 sov-schema-db = { path = "../sovereign-sdk/full-node/db/sov-schema-db" }
+sov-state = { path = "../sovereign-sdk/module-system/sov-state" }
 # sov-stf-runner = { path = "../sovereign-sdk/full-node/sov-stf-runner", features = ["native"] }
 
 # Citrea
 citrea-primitives = { path = "../primitives/" }
+citrea-stf = { path = "../citrea-stf" }
 citrea-storage-ops = { path = "../storage-ops" }
 
 [dev-dependencies]
