+++ crates/fullnode/src/lib.rs
@@ -4,6 +4,7 @@ use std::sync::Arc;
 use anyhow::Result;
 use citrea_common::backup::BackupManager;
 use citrea_common::cache::L1BlockCache;
+use citrea_common::l2::L2SyncWorker;
 use citrea_common::{InitParams, RollupPublicKeys, RunnerConfig};
 use citrea_stf::runtime::CitreaRuntime;
 use citrea_storage_ops::pruning::{Pruner, PrunerService};
@@ -17,7 +18,7 @@ use sov_modules_stf_blueprint::StfBlueprint;
 use sov_prover_storage_manager::ProverStorageManager;
 use sov_rollup_interface::services::da::DaService;
 use sov_rollup_interface::zk::ZkvmHost;
-use tokio::sync::{broadcast, Mutex};
+use tokio::sync::{broadcast, mpsc, Mutex};
 
 pub mod da_block_handler;
 pub mod db_migrations;
@@ -42,7 +43,8 @@ pub fn build_services<Da, DB, Vm>(
     code_commitments: HashMap<SpecId, <Vm as Zkvm>::CodeCommitment>,
     backup_manager: Arc<BackupManager>,
 ) -> Result<(
-    CitreaFullnode<Da, DB>,
+    CitreaFullnode<DB>,
+    L2SyncWorker<Da, DB>,
     L1BlockHandler<Vm, Da, DB>,
     Option<PrunerService>,
 )>
@@ -63,7 +65,8 @@ where
         PrunerService::new(pruner, last_pruned_block, soft_confirmation_tx.subscribe())
     });
 
-    let runner = CitreaFullnode::new(
+    let (l2_signal_tx, l2_signal_rx) = mpsc::channel(1);
+    let l2_sync_worker = L2SyncWorker::new(
         runner_config,
         init_params,
         native_stf,
@@ -74,7 +77,10 @@ where
         fork_manager,
         soft_confirmation_tx,
         backup_manager.clone(),
+        l2_signal_tx,
     )?;
+
+    let runner = CitreaFullnode::<DB>::new(ledger_db.clone(), l2_signal_rx)?;
 
     let l1_block_handler = L1BlockHandler::new(
         ledger_db,
@@ -87,5 +93,5 @@ where
         backup_manager,
     );
 
-    Ok((runner, l1_block_handler, pruner))
+    Ok((runner, l2_sync_worker, l1_block_handler, pruner))
 }
